+++ NightSlayer/Assets/_ProjectAssets/Scripts/Enemies/States/Data/D_RangedAttackState.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class D_RangedAttackState : ScriptableObject
 {
     public GameObject projectile;
-    public int projectileDamage = 10;
+    public float projectileDamage = 10f;
     public int knockbackForceX = 1;
     public int knockbackForceY = 1;
     public float projectileSpeed = 12f;
